Correct the source docs and centralize the example mDNS/CLI helpers
Refresh the internals and integration-guide prose that the one-atomic
lifecycle rework left stale: the write gate is the STREAM_ACCEPTING bit
of stream_state_ (not accepting_audio_), desired state is STREAM_DESIRED
(not stream_requested_), the host ring takes a short mutex so the guide
no longer calls the path universally callback-safe, and the ESP worker
allocation paragraph distinguishes the per-message text contexts from
the make_shared BinarySendLookup and its owner-scoped reclamation.
Extract the duplicated MdnsAdvertiser and log-level parser into
examples/common and share them across all three host examples.
