+++ examples/basic_client/main.cpp
@@ -36,6 +36,8 @@
 #include "sendspin/metadata_role.h"
 #include "sendspin/player_role.h"
 #ifdef SENDSPIN_HAS_PORTAUDIO
+#include "cli_util.h"
+#include "mdns_advertiser.h"
 #include "portaudio_sink.h"
 #endif
 
@@ -63,60 +65,6 @@ static const char* SENDSPIN_PATH = "/sendspin";
 // Tracks total audio bytes received (used when PortAudio is unavailable)
 static size_t null_audio_total_bytes = 0;
 
-#ifdef SENDSPIN_HAS_MDNS
-// Manages mDNS service advertisement via dns_sd.h
-class MdnsAdvertiser {
-public:
-    ~MdnsAdvertiser() {
-        stop();
-    }
-
-    bool start(const std::string& name, uint16_t port, const std::string& path) {
-        // Build TXT record with path and name keys
-        TXTRecordRef txt;
-        TXTRecordCreate(&txt, 0, nullptr);
-        TXTRecordSetValue(&txt, "path", static_cast<uint8_t>(path.size()), path.c_str());
-        TXTRecordSetValue(&txt, "name", static_cast<uint8_t>(name.size()), name.c_str());
-
-        DNSServiceErrorType err = DNSServiceRegister(
-            &service_ref_,
-            0,                    // flags
-            0,                    // interface index (0 = all)
-            name.c_str(),         // service name
-            "_sendspin._tcp",     // service type
-            nullptr,              // domain (default)
-            nullptr,              // host (default)
-            htons(port),          // port (network byte order)
-            TXTRecordGetLength(&txt),
-            TXTRecordGetBytesPtr(&txt),
-            nullptr,              // callback (not needed for simple registration)
-            nullptr               // context
-        );
-
-        TXTRecordDeallocate(&txt);
-
-        if (err != kDNSServiceErr_NoError) {
-            fprintf(stderr, "Failed to register mDNS service: error %d\n", err);
-            return false;
-        }
-
-        fprintf(stderr, "mDNS: Advertising _sendspin._tcp on port %u (name: %s)\n", port,
-                name.c_str());
-        return true;
-    }
-
-    void stop() {
-        if (service_ref_ != nullptr) {
-            DNSServiceRefDeallocate(service_ref_);
-            service_ref_ = nullptr;
-            fprintf(stderr, "mDNS: Service advertisement stopped\n");
-        }
-    }
-
-private:
-    DNSServiceRef service_ref_{nullptr};
-};
-#endif  // SENDSPIN_HAS_MDNS
 
 static std::atomic<bool> running{true};
 
@@ -136,15 +84,6 @@ static void print_usage(const char* prog) {
     fprintf(stderr, "  -h            Show this help\n");
 }
 
-static bool parse_log_level(const char* str, LogLevel& level) {
-    if (strcmp(str, "none") == 0) { level = LogLevel::NONE; return true; }
-    if (strcmp(str, "error") == 0) { level = LogLevel::ERROR; return true; }
-    if (strcmp(str, "warn") == 0) { level = LogLevel::WARN; return true; }
-    if (strcmp(str, "info") == 0) { level = LogLevel::INFO; return true; }
-    if (strcmp(str, "debug") == 0) { level = LogLevel::DEBUG; return true; }
-    if (strcmp(str, "verbose") == 0) { level = LogLevel::VERBOSE; return true; }
-    return false;
-}
 
 static bool parse_port(const char* str, uint16_t& port) {
     char* end = nullptr;
@@ -179,7 +118,7 @@ int main(int argc, char* argv[]) {
                 }
                 break;
             case 'l':
-                if (!parse_log_level(optarg, log_level)) {
+                if (!sendspin_examples::parse_log_level(optarg, log_level)) {
                     fprintf(stderr, "Unknown log level: %s\n", optarg);
                     print_usage(argv[0]);
                     return 1;
